backend/requirements.txt
fastapi
uvicorn
pyjwt
bcrypt
httpx
